Add --overwriteMode option to combine multiple studies.

+++ python/parse_data.py
@@ -7,6 +7,7 @@ import os
 import logging
 from datetime import datetime
 from csv import DictReader as CsvDictReader
+import re
 
 import pandas as pd
 from jsonschema import validate, ValidationError
@@ -34,10 +35,12 @@ SCHEMA = [
 '''
 CREATE TABLE replicates (
     replicateId INTEGER PRIMARY KEY,
-    replicate TEXT UNIQUE,
+    replicate TEXT NOT NULL,
+    project TEXT NOT NULL,
     acquiredTime BLOB NOT NULL,
     acquiredRank INTEGER NOT NULL,
-    ticArea REAL NOT NULL
+    ticArea REAL NOT NULL,
+    UNIQUE(replicate, project) ON CONFLICT FAIL
 )''',
 '''
 CREATE TABLE precursors (
@@ -66,6 +69,7 @@ CREATE TABLE metadata (
     replicateId INTEGER NOT NULL,
     annotationKey TEXT NOT NULL,
     annotationValue TEXT,
+    annotationType VARCHAR(6) CHECK( annotationType IN ('INT', 'FLOAT', 'BOOL', 'STRING')) NOT NULL DEFAULT 'STRING',
     FOREIGN KEY (replicateId) REFERENCES replicates(replicateId)
 )
 ''',
@@ -133,6 +137,17 @@ def _initialize(fname):
     return conn
 
 
+def pd_type_to_db_type(pdType):
+    if re.search('^int[0-9]+$', pdType):
+        return 'INT'
+    if re.search('^float[0-9]+$', pdType):
+        return 'FLOAT'
+    if pdType == 'bool':
+        return 'BOOL'
+    else:
+        return 'STRING'
+
+
 def read_metadata(fname, metadata_format=None):
     if metadata_format:
         _format = metadata_format
@@ -157,13 +172,16 @@ def read_metadata(fname, metadata_format=None):
 
     df = pd.DataFrame(data)
     df['Replicate'] = df['Replicate'].apply(lambda x: os.path.splitext(x)[0])
+    types = {k: pd_type_to_db_type(v.name) for k, v in df.dtypes.to_dict().items()}
     df = pd.melt(df, id_vars=['Replicate'], var_name='annotationKey', value_name='annotationValue',
                  value_vars=[x for x in list(df.columns) if x != 'Replicate'])
+    df['annotationType'] = df['annotationKey'].apply(lambda x: types[x])
 
     return df
 
 
-def write_db(fname, replicates, precursors, protein_quants=None, metadata=None):
+def write_db(fname, replicates, precursors, protein_quants=None, metadata=None,
+             projectName=None, overwriteMode='error'):
     '''
     Write reports to QC sqlite database.
 
@@ -179,6 +197,11 @@ def write_db(fname, replicates, precursors, protein_quants=None, metadata=None):
         Long formated protein quants dataframe (optional)
     metadata: pd.DataFrame
         Metadata dataframe (optional)
+    projectName: str
+        The project name to use in the replicates table.
+    overwriteMode: str
+        Behavior when the output file already exists.
+        Choices are 'error', 'overwrite', or 'append'
 
     Returns
     -------
@@ -186,50 +209,94 @@ def write_db(fname, replicates, precursors, protein_quants=None, metadata=None):
         True if all operations were sucessful, false otherwise.
     '''
 
+    conn = None
+    append = False
     if os.path.isfile(fname):
-        LOGGER.warning(f'{fname} already exists. Overwriting...')
-        os.remove(fname)
-
-    # make dict for replicateId column which links the replicate and precursor tables
-    repIndex = {r: i for i, r in zip(replicates['replicate'].index, replicates['replicate'])}
+        if overwriteMode == 'error':
+            LOGGER.error(f'{fname} already exists. Use the --overwriteMode option to append or overwrite')
+            return False
+        elif overwriteMode == 'overwrite':
+            LOGGER.warning(f'{fname} already exists. Overwriting...')
+            os.remove(fname)
+            conn = _initialize(fname)
+        elif overwriteMode == 'append':
+            conn = sqlite3.connect(fname)
+            append = True
+        else:
+            LOGGER.error(f'"{overwriteMode}" is an unknown overwriteMode!')
+            return False
+    else:
+        conn = _initialize(fname)
+            
 
     if protein_quants is not None:
         proteins = protein_quants[['name', 'accession', 'description']].drop_duplicates().reset_index(drop=True)
-        proteinIndex = {r: i for i, r in zip(proteins['accession'].index, proteins['accession'])}
+        proteinIndex = {r: i for i, r in zip(proteins['name'].index, proteins['name'])}
 
         # check that protein accessions in protein and precursor table match
-        if set(proteins['accession'].to_list()) != set(precursors['proteinAccession'].to_list()):
+        if set(proteins['name'].to_list()) != set(precursors['proteinName'].to_list()):
             LOGGER.error('Protein and precursor ProteinAccessions differ!')
             return False
-
-        protein_quants['replicateId'] = protein_quants['replicateName'].apply(lambda x: repIndex[x])
-        protein_quants['proteinId'] = protein_quants['accession'].apply(lambda x: proteinIndex[x])
-        protein_quants = protein_quants[['replicateId', 'proteinId', 'abundance']]
-
 
     else:
         proteins = precursors[['proteinName', 'proteinAccession']].drop_duplicates().reset_index(drop=True)
         proteins = proteins.rename(columns={'proteinName': 'name', 'proteinAccession': 'accession'})
         proteins['description'] = None
-        proteinIndex = {r: i for i, r in zip(proteins['accession'].index, proteins['accession'])}
+
+    if projectName is None:
+        projects = pd.read_sql('SELECT DISTINCT project FROM replicates;', conn)
+        projectName = f'project_{len(projects.index) + 1}'
+    replicates['project'] = projectName
+
+    # deal with existing proteins and replicates
+    if append:
+        conn = sqlite3.connect(fname)
+
+        # deal with existing replicates
+        curReplicates = pd.read_sql('SELECT * FROM replicates;', conn)
+        replicates.index = pd.RangeIndex(start=len(curReplicates.index), stop=len(curReplicates.index) + len(replicates.index), step=1)
+        repIndex = {r: i for i, r in zip(replicates['replicate'].index, replicates['replicate'])}
+
+        # deal with existing proteins
+        curProteins = pd.read_sql('SELECT * FROM proteins;', conn)
+        curProteinNames = set(curProteins['name'].to_list())
+        proteins = proteins[proteins['name'].apply(lambda x: x not in curProteinNames)]
+        proteins.index = pd.RangeIndex(start=len(curProteins.index), stop=len(curProteins.index) + len(proteins.index), step=1)
+        proteinIndex = {r: i for i, r in zip(curProteins['proteinId'], curProteins['name'])}
+        proteinIndex = {**proteinIndex, **{r: i for i, r in zip(proteins['name'].index, proteins['name'])}}
+
+    else:
+        # make dict for replicateId column which links the replicate and precursor tables
+        repIndex = {r: i for i, r in zip(replicates['replicate'].index, replicates['replicate'])}
+
+        # make dict for proteinId column which links the protein and proteinQuants and precursor tables
+        proteinIndex = {r: i for i, r in zip(proteins['name'].index, proteins['name'])}
 
     precursors['replicateId'] = precursors['replicateName'].apply(lambda x: repIndex[x])
-    precursors['proteinId'] = precursors['proteinAccession'].apply(lambda x: proteinIndex[x])
+    precursors['proteinId'] = precursors['proteinName'].apply(lambda x: proteinIndex[x])
     precursors = precursors[PRECURSOR_QUALITY_COLUMNS].drop_duplicates()
 
-    conn = _initialize(fname)
-    replicates.to_sql('replicates', conn, if_exists='append', index=True, index_label='replicateId')
+    if protein_quants is not None:
+        protein_quants['replicateId'] = protein_quants['replicateName'].apply(lambda x: repIndex[x])
+        protein_quants['proteinId'] = protein_quants['name'].apply(lambda x: proteinIndex[x])
+        protein_quants = protein_quants[['replicateId', 'proteinId', 'abundance']]
+
     proteins.to_sql('proteins', conn, if_exists='append', index=True, index_label='proteinId')
+    replicates.to_sql('replicates', conn, if_exists='append', index=True, index_label='replicateId')
     precursors.to_sql('precursors', conn, index=False, if_exists='append')
 
     if protein_quants is not None:
         protein_quants.to_sql('proteinQuants', conn, index=False, if_exists='append')
 
     if metadata is not None:
-        query='SELECT replicate, replicateId FROM replicates'
-        rep_ids = {fname: rep_id for fname, rep_id in conn.execute(query).fetchall()}
-        metadata['replicateId'] = metadata['Replicate'].apply(lambda x: rep_ids[x])
-        metadata = metadata[['replicateId', 'annotationKey', 'annotationValue']]
+        missing_metadata = [x for x in metadata['Replicate'].drop_duplicates().to_list() if x not in repIndex]
+        for rep in missing_metadata:
+            LOGGER.warning(f'Metadata row: \"{rep}\" does not exist in replicate report!')
+
+        metadata = metadata.loc[metadata['Replicate'].apply(lambda x: x in repIndex),]
+
+        metadata['replicateId'] = metadata['Replicate'].apply(lambda x: repIndex[x])
+        metadata = metadata[['replicateId', 'annotationKey', 'annotationValue', 'annotationType']]
         metadata.to_sql('metadata', conn, index=False, if_exists='append')
 
     conn.close()
@@ -248,6 +315,10 @@ def main():
                         help='Long formated protein abundance report.')
     parser.add_argument('-o', '--ofname', default='data.db3',
                         help='Output file name. Default is ./data.db3')
+    parser.add_argument('--overwriteMode', choices=['error', 'overwrite', 'append'], default='error',
+                        help='Behavior if output file already exists. '
+                             'By default the script will exit with an error if the file already exists.')
+    parser.add_argument('--projectName', default=None, help='Project name to use in replicates table.')
     parser.add_argument('replicates', help='Skyline replicate_report')
     parser.add_argument('precursors', help='Skyline precursor_report')
     args = parser.parse_args()
@@ -287,7 +358,9 @@ def main():
 
     # build database
     LOGGER.info('Writing database...')
-    if not write_db(args.ofname, replicates, precursors, protein_quants=protein_quants, metadata=metadata):
+    if not write_db(args.ofname, replicates, precursors,
+                    protein_quants=protein_quants, metadata=metadata,
+                    projectName=args.projectName, overwriteMode=args.overwriteMode):
         LOGGER.error('Failed to create database!')
         sys.exit(1)
     LOGGER.info('Done writing database...')
